🚧 Restructured shooter code for our 0 to 1 scale use
Essentially made it so from anywhere we access the shooter code, we only need to use this 0 to 1 scale. This will help us in the future if the absolute encoder's values are ever changed. Need to test, but I am pretty confident that it will work with minimal debugging. Code builds just fine.