🎉 New jetson branch with some code for week 1
Puts out what our current alliance is to the network table.
Tried to put it in the robot.h file but am running into issue, someone may be able to debug tn since im not there.
This is what we will use for week 1 because skew drive will not work for by then :/