✨ LEDs work on the arduino nano. I just have to fix somethings for the color sensor
I commented out the LEDs because the LEDs are using a for loop which is slowing down the communication of the Color senso. When I come back in I will find a way to do it without a color sensor or put the write function in the for loop, whichever one works better.